+++ src/Our.Umbraco.FriendlyRobots/Startup/RouteComposer.cs
@@ -1,7 +1,7 @@
 ﻿using Umbraco.Core;
 using Umbraco.Core.Composing;
 
-namespace Our.Umbraco.DynamicRobots.Startup
+namespace Our.Umbraco.FriendlyRobots.Startup
 {
     public class RouteComposer : IUserComposer
     {